Don't evaluate skip if include doesn't match
When include doesn't match we skip the file in any
case. Therefore we don't need to evaluate any of the
skip filters.

There was also this bug that when the file didn't match
include, but matched skip, it would be included twice
in the list of files to be skipped.